Color Picker has moved to its own component and tab switcher renders appropriate component
