+++ AwesomeProject/.expo/packager-info.json
@@ -2,8 +2,8 @@
   "devToolsPort": 19002,
   "expoServerPort": 19000,
   "packagerPort": 19001,
-  "packagerPid": 94712,
+  "packagerPid": 97011,
   "expoServerNgrokUrl": "https://va-t6p.anonymous.awesomeproject.exp.direct",
   "packagerNgrokUrl": "https://packager.va-t6p.anonymous.awesomeproject.exp.direct",
-  "ngrokPid": 94736
+  "ngrokPid": 97036
 }
